requirements.txt
joblib==0.14.1
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
